refactor pv-search, give board to evaluator

+++ src/types/piece.rs
@@ -86,7 +86,7 @@ impl From<u8> for PieceType {
 
 impl Step for Piece {
     fn steps_between(start: &Self, end: &Self) -> Option<usize> {
-        Some((*end as usize - *start as usize) as usize)
+        Some(*end as usize - *start as usize)
     }
 
     fn forward_checked(start: Self, count: usize) -> Option<Self> {
